feat: 프로젝트 목록 조회시 프로젝트 회원 응답 추가 (#31)
* feat: 프로젝트 초대 코드 생성, 목록 추가 기능 구현

* refactor: RedisService 동일 로직 메서드 공통화

* Revert "feat: 프로젝트 초대 기본 기능 구현 (#23)" (#24)

This reverts commit afbd02124e062e51a72ab68150c99ff8436140a0.

* feat: 프로젝트 회원 CRUD, 프로젝트 회원 초대 기능 추가

* feat: file 기본 세팅

* update

* feat: 프로젝트 목록 조회시 프로젝트 회원 목록 응답 추가

+++ src/main/java/com/douzone/prosync/project/dto/response/GetProjectResponse.java
@@ -14,7 +14,7 @@ import lombok.Getter;
 public class GetProjectResponse{
 
     @ApiModelProperty(example = "프로젝트 아이디")
-    private Integer projectId;
+    private Long projectId;
 
     @ApiModelProperty(example = "프로젝트 이름")
     private String name;
